util-docker: Simplify gcc and clang Dockerfiles
This works by building all GCC and Clang images on top of the 24.04
all-dependencies image. Fortunately 24.04 apt has all the gcc and clang
versions we support (assuming support for 20.04 is dropped).

This simplication is also more efficient. Each image only needs to `apt
install` the compiler (instead of all the dependencies needed, as
before).

Change-Id: Iefd3677a0e1e62ca86a6c92a616f39d4caef4c5b