cpu: Indirect predictor track conditional indirect
In the refactor made by commit f65df9b959d8e2fc3b5
conditional indirect branches are no longer
updated in the indirect predictor.

This kind of branches do not exist in x86 neither
arm, but they are present in PowerPC.

This patch, enables the indirect predictor to
track this kind of branches.

Change-Id: I16959095aba2e19677fde312e4e2345805c3e890