arch-arm: Fix 32-bit semihosting ABI
It appears we have been trying to read 64-bit arguments for ARM32 since
695583709b8c09d32f86057e63868bc125f1a86b. I noticed that SYS_OPEN
was trying to read a really long string as the pathname argument and it
turned out it was reading from the wrong stack offset. With this change
I can successfully run some of the semihosting tests for ARM32.

Change-Id: Ie154052dac4211993fb6c4c99d93990123c2eacf