arch-x86: break 32/64-bit mov's input dependency on prior dest value
Fix #1169. Break the input dependency of 32-bit and 64-bit 'mov'
micro-ops on the prior value in the destination register. Such a
dependency is required for 8-bit and 16-bit moves, as they do not
completely overwrite the value in the destination register. However,
it is unnecessary for 32-bit moves (which implicitly zero the upper
32 bits) and 64-bit moves.

This patch implements the fix by adding a new code template field inside
the generated constructors of X86StaticInst's, called `invalidate_srcs`,
which instruction implementations like `mov` can use to conditionally
invalidate particular source registers as needed.

Change-Id: Ib2aef6be6da08752640ea3414b90efb7965be924