cpu-o3: Fix partial store-to-load forwarding case handling
When a load instruction falls into partial store-to-load forwarding (STLF)
case, LSQUint marks the load to be reissued (Issued flag cleared) and
rescheduled during LSQUnit::read. In this case the instruction is not ready
for commit. But if the instruction encounters fault, e.g. by crossing page
boundary, the load instruction is erroneously marked Executed and sent to
commit stage, and in turn causes assertion failure when it's re-executed.
This commit fixes the problem by detecting the partial STLF and not sending
it to commit stage.

Change-Id: Ie481a5d1c97eaf62e0e92f2311b67c2afbe3c5d6