arch,cpu,sim: Add mechanism to partially print vector regs
Currently, gem5's inst tracer prints the whole vector register container
by default. The size of vector register containers in gem5 is the
maximum size allowed by the ISA. For vector-length agnostic (VLA) vector registers,
this means ARM SVE vector container is 2048 bits long, and RISC-V vector container
is 65535 bits long. Note that VLA implementation in gem5 allows the vector length
to be varied within the limit specified by the ISAs.

However, in most use cases of gem5, the vector length is much less than 65535 bits.
This causes two issues: (1) the vector container requires allocating and moving
around a large amount of unused data while only a fraction of it is used, and (2) printing
the execution trace of a vector register results in a wall of text with a small amount of
useful data.

This change addresses the problem (2) by providing a mechanism to limit the amount
data printed by the instruction tracer. This is done by adding an function printing
the first X bits of a vector register container, where X is the vector length determined
at runtime, as opposed to the vector container size, which is determined at compilation time.

Change-Id: I815fa5aa738373510afcfb0d544a5b19c40dc0c7
Signed-off-by: Hoa Nguyen <hn@hnpl.org>