arch: Add function to return VLA vector length for all ISAs
Vector-length Agnostic (VLA) is a style of vectorization in which the
vectorized code can run on any vector length implemented by the
microarchitecture. This vectorization style appears in SVE/SVE2 (an
extension of ARM ISA) and RVV (an extension of RISC-V ISA).

This change adds a function to the gem5's ISA interface returning
the vector length of the VLA registers. For ARM ISA, it returns
SVE_VL. For RISC-V ISA, it returns VLEN. For other ISAs, it returns -1.

Change-Id: I8e622c9ab47a36770479c0ff68a15522602c82bf
Signed-off-by: Hoa Nguyen <hn@hnpl.org>