sim: Correct DataStatus set by setData() when RegClass is known
DataStatus is used by InstTracer to determine the data format of the content
of the destination register of an instruction. The types consists of integer
types (`DataInt*`), a floating point type (`DataDouble`), and a vector type
(`DataReg`).

Currently, the `setData(RegClass, *)` function assumes the register value
to be of type `DataReg`, which means the content of the register is an
array of values, for every `RegClass`. However, there are cases in the ISA
implementation that the `setData` function above is used for writing an
integer to the trace.

This change addresses this issue by setting DataStatus accordingly to RegClass.

Change-Id: I79423a4942ab2a3fde5c9cf86de0d1fced648cf0
Signed-off-by: Hoa Nguyen <hn@hnpl.org>