cpu, stdlib: Fixing access to index in SpatterGen.
This change fixes the way indices are generated in a multi
generator setup. It changes it from all cores generating the same
trace of indices for accessing the index array to each core
generating an interleaved subset of indices. For an example look
below for traces (indices to index array) in a 2 core setup.
Before:
core_0: 0, 1, 2, 3, 4, 5, 6, 7, ...
core_1: 0, 1, 2, 3, 4, 5, 6, 7, ...
After:
core_0: 0, 1, 2, 3, 8, 9, 10, 11, ...
core_1: 4, 5, 6, 7, 12, 13, 14, 15, ...

Additionally, this change fixes the SpatterKernel class in
the standard library to comply with the change in the
SpatterGen source code.

Change-Id: Idae69486eafafa21aaaad1239944c34ba8092087