arch-riscv: Fix TLB lookup with vaddrs
Previously, all of the TLB lookup/insert functions were using the
full virtual addresses even though the variables in the functions
said "vpn." This change explicitly converts the virtual address
to the VPN without any least significant zeros for the offset.
I.e., vpn >> page_size.

The main bug solved in this changeset is the asid was |'d with
the upper bits of the virtual address, but sometimes there were
all 1's. Therefore, you could get a TLB hit even if the ASID was
different. Interestingly, the page that seemed to cause these
issues was a 1 GiB page.

This change also starts refactoring some of the page table
details to support sv46 and sv57 page table formats.

In my testing, the Linux kernel boot uses large pages (even
OpenSBI uses large pages), so it seems that large pages also
work. However, this seems like magic to me, so I'm not sure if
it's correct.

This change also updates some asserts, and debug statements with
more useful debugging information.

Partially fixes #1235. More testing needs to be done to be
confident.

Change-Id: If832fe8ae89874cf48d7367f5694c5adba2bd6d7
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>