gpu-compute,tests: Move GPU tests to testlib
A new host tag `gcn_gpu` has been added. This allows for selection of
those GPU tests which depend upon the gcn-gpu docker image to run.

In addition to this, the square GPU tests has been moved to the CI
tests. This ensures some GPU code is compiled and run on every PR.

Change-Id: Ic1a3e3a5c14566c5a53aeb74ab6d0725a1310fff