arch-riscv: use sign-extend for all address generation
In gem5, we use the same code base for RISC-V 32 and 64.

However, if we need to allow modifiable XLEN control on CSR.mstatus in
the future, we should follow the RISC-V ISA manual to sign-extend
all the register results, including PC and GPR. If this feature
implemented, the simulator needs to handle user-mode in RV32 but
CSR.SATP sets to Sv39. In this case, 0x80000000 and 0xffffffff80000000
are different addresses in the 64-bit S-Mode perspective, but they
are the same in the 32-bit U-Mode perspective. We should avoid this
wrong behavior happening before we implement this feature.

Thus, we need to sign-extend the results of all the addresses,
including the PC and memory addresses, which currently use zero-extend.
As specified in the RISC-V ISA manual, we use zero-extend in narrow
XLEN mode for the physical address implemented in TLB.

Changes based on spec:
1. Sign-extend narrow XLEN: https://github.com/riscv/riscv-isa-manual/blob/riscv-isa-release-b7a445a-2024-07-02/src/machine.adoc?plain=1#L567
2. Zero-extend physical address: https://github.com/riscv/riscv-isa-manual/blob/riscv-isa-release-b7a445a-2024-07-02/src/supervisor.adoc?plain=1#L1670

Change-Id: I01640fcd10ec5a2ff7289765553cdc047f3502c6
Signed-off-by: Yangyu Chen <cyy@cyyself.name>