arch-arm: Fix incorrect operation of VRINT* instructions
After a lot of debugging and comparing traces I noticed that vrintp
was giving different results from QEMU. An input of 0x3f800000 (1.0)
was being passed to the fplib helpers as (uint32_t)1 which has a
completely different floating-point interpretation and the result was
therefore completely wrong.

I've fixed this as well as all remaining implicit float-to-int
conversions in the ARM instruction execution. There are more
-W(implicit-)float-conversion warnings in the other executors, but for
now this fixes the issue I was seeing.

Change-Id: Ifdeee745ca155d7f4504ac4c54235ac431acdeb9