sim: Add error message for kernel exceeding memory size
This commit adds an error message to src/sim/kernel_workload.cc
to tell the user to increase the memory size of their configuration
when the size of the kernel is greater than the size of memory.
The error message also specifies the minimum memory size needed
to fit the kernel.

Change-Id: I7d8f50889ed8172f64b84f98301a35e5f2f352d3