stdlib: Adding Modification support in stdlib.
This implements a capability in AbstractBoard to allow running
user defined code between calling _pre_instantiate and
m5.instantiate while using the Simulator pacakge. The simulator
object will call a secondary function from the board called
`_apply_modifiers` to allow user provided function to be
executed after everything is connected.

This change adds the following:
- A step in the instantiation process in the Simulator class that
calls the `_apply_modifiers` method from the board passed to the
Simulator object.
- `_apply_modifier` method to the AbstractBoard that does not
really require further extension that calls on the modifiers
attached to the board to apply their modification.
- `_modifiers` attribute to the AbstractBoard to store all
the modifiers attached to the board.
- `add_modifier` method to add a modifier object to the
`_modifiers` attribute.
- Example modifier for modifying the branch predictor in an
O3CPU core.

This is useful since it makes fast and reusable changes to the
parameters of a component in the standard library. An important
use case for this would be to explore the parameter space of
different SimObjects in gem5 for building reliable prebuilts.

Change-Id: I32ec44bbbef80c878feeb6edc5245ac3648479e5