arch-riscv: Overwrite getEMI() for timing expr
TimingExpression enables runtime calculation of the commit latency
in MinorCPU. For this, machInst is obtained by getEMI() to match it
with a given instruction. At default, getEMI() always returns 0 and is
therefore overwritten to enable timing expressions for RISC-V. This was
already done for ARM (see src/arch/arm/insts/static_inst.hh).

Change-Id: I03d669b3439fd24e00cbf893f5db9951dfe56b1f
Signed-off-by: Robert Hauser <robert.hauser@uni-rostock.de>