sim: Ensure probe pointers are connected to compatible listeners
I was trying to add PMU events for caches and noticed that we were
calling the notify() function with a nonsense increment since cache
events pass a CacheAccessProbeArg& to notify() and the RegularProbe in
the Arm PMU expects a uint64_t&. The result was that the counters was
incremented by the value of the PacketPtr member in the
CacheAccessProbeArg instead of incrementing by one. To avoid similar
difficult to debug issues, use a dynamic_cast<> in addListener().
This assertion is possible now that we no longer pass partially
constructed objects to this function.

Change-Id: I4aba443b46df5fbbe41d5fb6f9b2fe424d9f806e