misc,tests: Rm gem5 binary pre-build from dailys
Pre-building the gem5 binaries and having the jobs download them
when needed has proven to be too much for our runners to download.
This patch has each job compile its own gem5 binaries.
The goal of this patch is to get the daily tests passing again.
This prebuilding approach can be returned to later when and
if a working solution is found.

Change-Id: I3ac1469f9ccdf291fe1e67f07ed7474df4145fac