cpu-o3: Panic if no FU exists for an instruction needing to issue
If an instruction requires a functional unit that is not present
in the model (e.g., because it is not present in the configuration),
O3CPU treats it as a 1-cycle operation.

This commit changes the behavior to make the cpu panic when this
happens. The cpu panics only if the instruction reaches the
head of the ROB, meaning it is ok to have unsupported
instructions on the wrong path.

Thanks to Chandana S. Deshpande (deshpande.s.chandana@gmail.com) for
finding the issue.

Change-Id: Icd86041262dcdef1d3ce096da9855ac452b9c440