arch-riscv: Fix GDB breakpoint addr issue after #1316
The PR https://github.com/gem5/gem5/pull/1316 changes to set pc with
sign-extend address in order to support run RV32 program in RV64
architecture. However, it will let GDB with inconsistent address with
gem5 PCState (0x80001000 in GDB and 0xffffffff80001000 in gem5 PCState).

In the PR, we will start define the difference rvType in PCState or ISA
and machInst.rv_type. The rvType should define which XLEN is used for
architecture, and machInst.rv_type is used to define XLEN from current
privilege model (xstatus.sxl in supervisor, and xstatus.uxl in user).

Bug: 360011842
Test: None
Change-Id: Iac56b47b89e2402f1af3a12583c08d5078c7b677