sim-se: Replaces std::list with std::multiset for _vmaList
Several operations (mapRegion/UnmapRegion/fixupFault) involving the
_vmaList had time complexity O(n), where n is the number of currently
mapped virtual memory pages. This was because the list of VMA objects
had to be linearly iterated for each search.

These operations could be sped up to O(logn) using a full-fledged
Interval Tree, but std::multiset also does the job. This is because for
any VMA object in _vmaList, (1) its AddrRange is never interleaved
(AddrRange::_masks.size()==0) and (2) its AddrRange never overlaps with
that of any other VMA object in the list (MemState::mapRegion() avoids
creating a region that has preexisting mappings). This means a strict
weak ordering of VMA objects, which is necessary for std::multiset to
behave properly, can be established by simply comparing their start and
end addresses in the current use case of _vmaList.

This PR assumes MemState::remapRegion() never remaps a region to a
preexisting mapping (similar to MemState::mapRegion()).

If VMA objects in the _vmaList have an AddrRange that is interleaved or
overlaps with each other's intervals, then this PR becomes incorrect.

Change-Id: Iabfc9fafefca93de9788a4c0eae983b5de9847ae