stdlib: Add RiscvDemoBoard
This commit adds a RiscvDemoBoard that can be used with both SE and
FS mode. The commented out parts of the board configuration are from
the X86DemoBoard while the uncommented parts are from
configs/example/gem5_library/riscv-ubuntu-run.py. This was tested
using the workloads riscv-matrix-multiply-run for SE and
riscv-ubuntu-20.04-boot for FS.

Change-Id: I5d338562f92c27a89101d6a8e70090e91faafe6f