stdlib: Use board get_mem_ports consistently
Previously, whether the board object or the memory_system returned
the memory ports was not consistent in the cache_hierarchies

This commit makes it consistently use the board. Note: the board
is a better place so it can customize the ports (e.g., add I/O
components or other things.

This commit also makes the arm board consistent with the other
boards and removes the specialized `get_mem_ports` that was not
used.

Change-Id: Ibbd5291ae8e0fee10aefdbd8890a653b8bd737df
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>