mem-ruby: Update GPU_VIPER to support WB L2 in dGPUs
Previously, when SDMAs read data from dGPU, the requests went directly
to memory. If the GPU is configured such that the L2 is write-back in
nature, then the SDMA returns incorrect data from memory since the most
recently updated version of the data would be in L2. This update
modifies the cache coherence protocol to probe L2 for data before
sending requests to memory.

Change-Id: I874e838fda2862aa7d67094de0520272c9a7cdce