arch-riscv: Fix CLINT mtime reset handling
The previous https://github.com/gem5/gem5/pull/1617 introduce
the CLINT reset feature. When reset, we changed the mtime to 0
and keep mtimecmp unchanged by default, we also need to check
mtime & mtimecmp regiter to update the MTI signal. However,
the mtime register will be incremented to 1 by `raiseInterruptPin`.

In the PR, we introduced the interrupt ID for CLINT, the mtime
will be incremented only if received the RTC signal