tests: Add tests for checking that stats reset
This commit adds tests to check that stats reset. The first of
these tests resets the stats before dumping them and checks that
the stats have values of 0 or nan. The second test runs a
simulation for 5 billion ticks, dumps the stats, then resets
and runs the simulation for another 10 billion ticks and dumps the
stats. (Most of) the values from these runs are added together
and compared against a run that runs for 15 billion ticks without
resetting.
There are a few statistics that do not appear to reset, but I am
not sure what the expected behavior for these is. For now, they
are not checked when the tests are run, but I have indicated
which stats these are in the test scripts.