cpu: BTB update at commit option
Option to update the BTB at either squash or commit time.
Updating at squash can be useful to update the BTB earlier
and avoid BTB misses on subsequent branches.
However, it can also lead to BTB pollution if the branch
is on the false path and will be squashed later.
To the best of my knowledge the default should be at commit.
Furthermore, the statistics and probe point for conditional
branches was moved to commit.

Signed-off-by: David Schall <david.schall@tum.de>