dev,arch-riscv: Fix Clint msip register read/write
From SiFive U54MC datasheet

```
Each msip register is a 32-bit wide WARL register, where the upper 31 bits
are tied to 0. The least-significant bit is reflected in the MSIP bit of
the mip CSR. Other bits in the msip register are hardwired to zero.
```

The register needs to clear upper 31 bits rather than rely on an assertion
check. For register reader part, we don't need to sync with mip.MSIP
because the bit is read-only.

Change-Id: I9e52c039d4c689b757e85644a055ae341b9ba0bb