scons: use --git-common-dir to detect hooks
To detect if pre-commit hooks are installed during compilation, the
command `git rev-parse --git-dir` is used which return the path to the
.git directory of current repository.

However, when using `git worktree` the above command returns for
linked-worktree:

`<original repo>/.git/worktrees/<name of worktree>`

But hooks are shared between all worktree and situated in
`<original repo>.git/hooks/` and so, scons will always ask to add
pre-commit hooks on compilation even if they are already installed.

Using `--git-common-dir` instead of `--git-common` will return the path
to `<original repo>/.git` which patch that bug.